Rename client module to lsp_client
